Make time a real plugin
