+++ zmon_worker_monitor/builtins/plugins/time_.worker_plugin
@@ -1,0 +1,17 @@
+[Core]
+Name = time
+Module = time_
+
+[Documentation]
+Author = John Doe
+Version = 0.1
+Website = https://github.com/zalando/zmon-worker
+Description = Builtin plugin that provides an time functions.
+
+
+[Configuration]
+
+;; The configuration section is optional, you can put here any parameters needed to configure the plugin.
+;; Note that these values may are overridden by zmon's global config.
+;; key = valueX
+;; Note valueX will be overridden if Zmon's global config has: plugin.{plugin_name}.configuration.{key} = valueY
