+++ zmon_worker_monitor/zmon_worker/common/time_.py
@@ -88,40 +88,3 @@ def parse_datetime(s):
         except:
             pass
     return None
-
-
-class TimeWrapper(object):
-
-    def __init__(self, spec='now', utc=False):
-        now = (datetime.utcnow() if utc else datetime.now())
-        delta = parse_timedelta(spec)
-        if delta:
-            self.time = now + delta
-        elif spec == 'now':
-            self.time = now
-        else:
-            self.time = parse_datetime(spec)
-
-    def __sub__(self, other):
-        '''
-        >>> TimeWrapper('2014-01-01 01:01:25') - TimeWrapper('2014-01-01 01:01:01')
-        24.0
-        '''
-
-        return (self.time - other.time).total_seconds()
-
-    def isoformat(self, sep=' '):
-        return self.time.isoformat(sep)
-
-    def format(self, fmt):
-        '''
-        >>> TimeWrapper('2014-01-01 01:01').format('%Y-%m-%d')
-        '2014-01-01'
-
-        >>> TimeWrapper('-1m').format('%Y-%m-%d')[:3]
-        '201'
-        '''
-
-        return self.time.strftime(fmt)
-
-
