+++ zmon_worker_monitor/zmon_worker/functions/__init__.py
@@ -2,11 +2,9 @@
 # -*- coding: utf-8 -*-
 
 from sql import SqlWrapper
-from time_ import TimeWrapper
 from zmon import ZmonWrapper
 
 __all__ = [
     'SqlWrapper',
-    'TimeWrapper',
     'ZmonWrapper'
 ]
